Rename bcp.ts -> connection.ts
