+++ packages/iov-bcp-types/src/atomicswap.ts
@@ -1,6 +1,6 @@
 import { Stream } from "xstream";
 
-import { BcpCoin, BcpConnection, BcpQueryEnvelope } from "./bcp";
+import { BcpCoin, BcpConnection, BcpQueryEnvelope } from "./connection";
 import { Address } from "./signables";
 import { SwapIdBytes } from "./transactions";
