+++ packages/iov-bcp-types/src/connection.ts
