+++ packages/iov-bcp-types/types/connection.d.ts
